Rest & App: GraphQL Stuff

+++ rest-expressjs/src/graphql/resolvers.ts
@@ -1,4 +1,6 @@
 import { Memo } from "../db/mongoose.controller.js";
+import validator from "validator";
+import type { AuthRequest } from "../utils/types.js";
 
 export const resolvers = {
   getHelloWorld: () => {
@@ -7,19 +9,47 @@ export const resolvers = {
       altText: "Hello from GraphQL!",
     };
   },
-  getMemos: async () => {
+  getMemos: async (_: any, req: AuthRequest) => {
+    if (!req.userId) {
+      throw new Error(
+        (req.errorMessage || "Unauthorized") + ": " + (req.errorStatus || 403)
+      );
+    }
     return await Memo.find();
   },
-  addMemo: async ({ input }: { input: InstanceType<typeof Memo> }) => {
+  addMemo: async (
+    { input }: { input: InstanceType<typeof Memo> },
+    req: AuthRequest
+  ) => {
+    if (!req.userId) {
+      throw new Error(
+        (req.errorMessage || "Unauthorized") + ": " + (req.errorStatus || 403)
+      );
+    }
     // "Validation" :D
-    if (input.title === "error") {
-      throw new Error("Title cannot be 'error'");
+    const errors = [];
+
+    if (!validator.isLength(input.title, { min: 5 })) {
+      errors.push("Title must be at least 5 characters long");
+    }
+
+    if (validator.equals(input.title, "error")) {
+      errors.push("Title cannot be the word 'error'");
+    }
+
+    if (errors.length > 0) {
+      throw new Error(errors.join(", "));
     }
 
     const newMemo = new Memo(input);
     return await newMemo.save();
   },
-  deleteMemo: async ({ id }: { id: string }) => {
+  deleteMemo: async ({ id }: { id: string }, req: AuthRequest) => {
+    if (!req.userId) {
+      throw new Error(
+        (req.errorMessage || "Unauthorized") + ": " + (req.errorStatus || 403)
+      );
+    }
     const deletedMemo = await Memo.findByIdAndDelete(id);
     return deletedMemo;
   },
